Move files into minimum-order-card folder.
