+++ js/src/components/flat-shipping-rates-input-cards/minimum-order-card/index.js
@@ -1,0 +1,1 @@
+export default './minimum-order-card';
