+++ js/src/components/flat-shipping-rates-input-cards/minimum-order-card/minimum-order-input-label-text.js
